Define SOCIAL_BAR_IS_VISIBLE_EVENT locally.
EosMetrics.EVENT_SOCIAL_BAR_IS_VISIBLE is deprecated because event IDs
should be defined in the file in which they are used. Otherwise we run
into problems when changing the wire protocol of those event IDs since
the event ID needs to change in lock step with the auxiliary payload.

[endlessm/eos-sdk#3029]