Delay the creation of WMInspect until the sidebar has been made visible
Apparently, there's some weird race condition that sometimes causes
_onActiveWindowChanged() to be called with the wrong window, which
makes it fail all the checks to see if the XID belongs to either the
sidebar or a transient window, finally calling to gtk_widget_hide().

The net result of this is that the shell and the social bar go somehow
out-of-sync with regard to the visibility of the sidebar, which ends
up never being shown on screen.

I've tried hard to find a fix but couldn't, so as a workaround, simply
wait until the window is visible before creating the WMInspect object
and connecting to the signals reporting a change in the active window.

https://phabricator.endlessm.com/T10999