socialBarView: Fix the delayed creation of WMInspect
Using state flags changed does not work around the race reliably - in
my VM more often than not, the window does not show up on the first
click on the social bar button.

Create the WMInspect instance right after the window is mapped. We can
use the map-event for this.

I needed to make the similar fix in the coding game manager, so please
see the ticket for some logs that explain the race.

https://phabricator.endlessm.com/T16113